requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.26.0
config==0.5.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap4==3.0.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-registration==2.4.1
django-utils-six==2.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.8.6
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
